pyproject.toml
[tool.poetry]
name = "luzia-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "luzia_api"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
matplotlib = "^3.5.2"
pandas = "^1.4.3"
opencv-python = "^4.6.0.66"
scikit-image = "^0.19.3"
scikit-learn = "^1.1.2"
tensorflow-gpu = "2.7.0"
python-dotenv = "^0.20.0"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
pillow = "^8.4.0"
numpy = "^1.21.2"
mysql-connector-python = "^8.0.27"
google-cloud-storage = "^2.8.0"
python-multipart = "^0.0.6"
ipython = "^8.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
